+++ src/vscode/index.ts
@@ -1,3 +1,3 @@
-export * from './vscode';
-
+export * from './devcontainer';
 export * from './launch-config';
+export * from './vscode';
